Implemented sequelize version of previous pg driver code
Takes up more space in database.js by defining models, however the authenticate() command has no equivalent in pg driver AFAIK, so if it disconnects there I can't check if the connection is there or to reestablish it. Also the automatic updating of created_at and updated_at might come useful in the future. Still debating whether or not to use the pg driver or sequelize