/team [teamname] now works using teamDictionary, teamprofiles and roster postgresql database
Checks, inserts & updates all happen automatically. Needs refactoring to avoid duplicate code / to make it neater. Perhaps reverting to raw pg driver would make this easier (especially with the freedom of creating your own queries, this can still be done in sequelize so that might be better to retain the automatic timestamps and authenticate command)