More elegant database solution for /team and /teamstats
Removed the two way conversion, instead one way conversion and changing the slash command hltv functions (not the API calls) to accept the converted values. Allows better cohesiveness between the database and the backend visual display to the user. /team and /teamstats are functioning in this way. /teammaps is partially implemented however in the old methodology so needs to be changed similarly to /teamstats and /team. Also checkTeamStats and checkTeamMaps need to be called from the other command as they share a HLTV API call to populate the database more reliably and avoids further expensive HLTV.getTeamStats API calls.